all errors container is added
